+++ flumine/config.py
@@ -5,9 +5,11 @@ simulated = False
 
 instance_id = None  # instance id (e.g. AWS ec2 instanceId)
 
-hostname = socket.gethostname()[
+customer_strategy_ref = socket.gethostname()[
     :15
 ]  # ie. docker container id (used as order customerStrategyRefs)
+
+hostname = customer_strategy_ref
 
 process_id = os.getpid()  # process id of app