+++ tests/test_streams.py
@@ -1,5 +1,6 @@
 import unittest
 from unittest import mock
+from unittest.mock import call
 
 from flumine.streams import streams, datastream, historicalstream
 from flumine.streams.basestream import BaseStream
@@ -57,8 +58,11 @@ class StreamsTest(unittest.TestCase):
         )
         self.assertEqual(len(mock_strategy.streams), 1)
         self.assertEqual(len(mock_strategy.historic_stream_ids), 1)
-        mock_get_file_md.assert_called_with(
-            "dubs of the mad skint and british", "marketType"
+        mock_get_file_md.assert_has_calls(
+            [
+                call("dubs of the mad skint and british", "marketType"),
+                call("dubs of the mad skint and british", "countryCode"),
+            ]
         )
 
     @mock.patch("flumine.streams.streams.get_file_md", return_value="PLACE")
@@ -81,8 +85,38 @@ class StreamsTest(unittest.TestCase):
         mock_add_historical_stream.assert_not_called()
         self.assertEqual(len(mock_strategy.streams), 0)
         self.assertEqual(len(mock_strategy.historic_stream_ids), 0)
-        mock_get_file_md.assert_called_with(
-            "dubs of the mad skint and british", "marketType"
+        mock_get_file_md.assert_has_calls(
+            [
+                call("dubs of the mad skint and british", "marketType"),
+                call("dubs of the mad skint and british", "countryCode"),
+            ]
+        )
+
+    @mock.patch("flumine.streams.streams.get_file_md", return_value="PLACE")
+    @mock.patch("flumine.streams.streams.Streams.add_historical_stream")
+    def test_call_backtest_country_code(
+        self, mock_add_historical_stream, mock_get_file_md
+    ):
+        self.mock_flumine.BACKTEST = True
+        mock_strategy = mock.Mock(
+            streams=[],
+            historic_stream_ids=[],
+            market_filter={
+                "markets": ["dubs of the mad skint and british"],
+                "listener_kwargs": {"canary_yellow": True},
+                "country_codes": ["VERDANSK"],
+            },
+        )
+        self.streams(mock_strategy)
+
+        mock_add_historical_stream.assert_not_called()
+        self.assertEqual(len(mock_strategy.streams), 0)
+        self.assertEqual(len(mock_strategy.historic_stream_ids), 0)
+        mock_get_file_md.assert_has_calls(
+            [
+                call("dubs of the mad skint and british", "marketType"),
+                call("dubs of the mad skint and british", "countryCode"),
+            ]
         )
 
     @mock.patch("flumine.streams.streams.Streams.add_historical_stream")